pyproject.toml
[tool.poetry]
name = "books-drf"
version = "0.1.0"
description = ""
authors = ["Yauheni_Liubimov <0.5TELOO@gmail.com>"]
readme = "README.md"
packages = [{include = "books_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"
black = "^23.1.0"
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py310"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.artifacts.*
    | .*\.github.*
    | .*\.idea.*
    | .*\.vscode.*
    | .*docs.*
)
'''
